refactor: extract segment usage read model
