+++ src/lib/features/change-request-access-service/fake-change-request-access-read-model.ts
@@ -34,8 +34,4 @@ export class FakeChangeRequestAccessReadModel
     public async isChangeRequestsEnabledForProject(): Promise<boolean> {
         return this.isChangeRequestEnabled;
     }
-
-    public async isSegmentUsedInActiveChangeRequests(): Promise<boolean> {
-        return this.isSegmentUsedInActiveChangeRequestsValue;
-    }
 }
