+++ src/lib/features/change-request-access-service/sql-change-request-access-read-model.ts
@@ -80,25 +80,4 @@ export class ChangeRequestAccessReadModel
         const { present } = result.rows[0];
         return present;
     }
-
-    public async isSegmentUsedInActiveChangeRequests(
-        segmentId: number,
-    ): Promise<boolean> {
-        const result = await this.db.raw(
-            `SELECT *
-            FROM change_request_events
-            WHERE change_request_id IN (
-            SELECT id
-            FROM change_requests
-            WHERE state IN ('Draft', 'In Review', 'Scheduled', 'Approved')
-            )
-            AND action IN ('updateStrategy', 'addStrategy');`,
-        );
-
-        const isUsed = result.rows.some((row) =>
-            row.payload?.segments?.includes(segmentId),
-        );
-
-        return isUsed;
-    }
 }
