+++ src/lib/features/change-request-segment-usage-read-model.ts/change-request-segment-usage-read-model.ts
@@ -1,0 +1,3 @@
+export interface IChangeRequestSegmentUsageReadModel {
+    isSegmentUsedInActiveChangeRequests(segmentId: number): Promise<boolean>;
+}
